Factored mysql.migrations out into a separate lib/migrations.js file.
