+++ lib/mysql.js
@@ -1,65 +1,15 @@
-var path = require('path');
 var mysql = require('mysql');
 var async = require('async');
-var migrate = require('db-migrate');
 var conf = require('../lib/configuration').get('database');
 var client = mysql.createClient(conf);
 var testDb = "`test_" + conf.database + "`";
 var usingTestDb = false;
 var dbEncoding = 'utf8';
 
-exports.migrations = {
-  dir: path.resolve(__dirname, '..', 'migrations'),
-  up: function(options, callback) {
-    if (arguments.length == 1) {
-      callback = options;
-      options = {};
-    }
-    options.config = options.config || exports.getDatabaseConfig();
-    options.count = options.count || Number.MAX_VALUE;
-    migrate.connect(options.config, function(err, migrator) {
-      if (err) throw err;
-      migrator.migrationsDir = exports.migrations.dir;
-      migrator.driver.createMigrationsTable(function(err) {
-        if (err) throw err;
-        migrator.up(options, function(err) {
-          migrator.driver.close();
-          callback(err);
-        });
-      });
-    });
-  },
-  down: function(options, callback) {
-    if (arguments.length == 1) {
-      callback = options;
-      options = {};
-    }
-    options.config = options.config || exports.getDatabaseConfig();
-    if (!options.destination && !options.count)
-      options.count = 1;
-    migrate.connect(options.config, function(err, migrator) {
-      if (err) throw err;
-      migrator.migrationsDir = exports.migrations.dir;
-      migrator.driver.createMigrationsTable(function(err) {
-        if (err) throw err;
-        migrator.down(options, function(err) {
-          migrator.driver.close();
-          callback(err);
-        });
-      });
-    });
-  }
-};
-
 exports.getDatabaseConfig = function() {
   var config = JSON.parse(JSON.stringify(conf));
   if (usingTestDb) config.database = testDb.slice(1, -1);
   return config;
-};
-
-exports.createTables = function (callback) {
-  callback = callback || function(){};
-  exports.migrations.up(callback);
 };
 
 exports.createTestDatabase = function (callback) {
